# backend/RestApiSwagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger PSWTM - OpenAPI 3.0",
    "description": "REST API Methods description for our Project",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:8800/api/v1"
    }
  ],
...
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "POST users as admin (Mitarbeiter Verwalten)",
        "description": "This can only be done by the logged in admin user.",
        "operationId": "createUser",
        "requestBody": {
          "description": "Created user object",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "James"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@email.com"
                  },
                  "is_admin": {
                    "type": "integer",
                    "description": "User Admin (1 yes, 0 no)",
                    "format": "int32",
                    "example": "false"
                  }
                },
                "xml": {
                  "name": "user"
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1
                    },
                    "username": {
                      "type": "string",
                      "example": "John.James"
                    },
                    "password": {
                      "type": "string",
                      "example": "$2a$08$kldeLiMkVHZpg/UkJUnaROFLrxVy7Vah9.WTQOtSa7KxGNfauA3R."
                    },
                    "firstName": {
                      "type": "string",
                      "example": "John"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "James"
                    },
                    "email": {
                      "type": "string",
                      "example": "john@email.com"
                    },
                    "credits": {
                      "type": "number",
                      "format": "float",
                      "example": "0"
                    },
                    "isAdmin": {
                      "type": "integer",
                      "description": "User Admin (1 yes, 0 no)",
                      "format": "int32",
                      "example": "false"
                    }
                  },
                  "xml": {
                    "name": "user"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "users"
        ],
        "summary": "GET user by username (Mitarbeiter Verwalten)",
        "description": "Can only used by admin",
        "operationId": "getUserByName",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "The name that needs to be fetched. Use user1 for testing. ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1
                    },
                    "username": {
                      "type": "string",
                      "example": "John.James"
                    },
                    "password": {
                      "type": "string",
                      "example": "$2a$08$kldeLiMkVHZpg/UkJUnaROFLrxVy7Vah9.WTQOtSa7KxGNfauA3R."
                    },
                    "firstName": {
                      "type": "string",
                      "example": "John"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "James"
                    },
                    "email": {
                      "type": "string",
                      "example": "john@email.com"
                    },
                    "credits": {
                      "type": "number",
                      "format": "float",
                      "example": "0"
                    },
                    "isAdmin": {
                      "type": "integer",
                      "description": "User Admin (1 yes, 0 no)",
                      "format": "int32",
                      "example": "false"
                    }
                  },
                  "xml": {
                    "name": "user"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/users/credit/{user_id}": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "PUT existing userCredit (Mitarbeiter Verwalten)",
        "description": "This can only be done by the logged in user admin.",
        "operationId": "updateUserCredit",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "name that need to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update an existent user in the store",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                  },
                  "username": {
                    "type": "string",
                    "example": "theUser"
                  },
                  "password": {
                    "type": "string",
                    "example": "12345"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "James"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@email.com"
                  },
                  "isAdmin": {
                    "type": "integer",
                    "description": "User Admin (1 yes, 0 no)",
                    "format": "int32",
                    "example": 1
                  }
                },
                "xml": {
                  "name": "user"
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "GET user by id (Mitarbeiter Verwalten)",
        "description": "Can only used by admin",
        "operationId": "getUserByID",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "The user_id that needs to be fetched. Use 1 for testing. ",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer",
                      "format": "int64",
                      "example": 1
                    },
                    "username": {
                      "type": "string",
                      "example": "John.James"
                    },
                    "password": {
                      "type": "string",
                      "example": "$2a$08$kldeLiMkVHZpg/UkJUnaROFLrxVy7Vah9.WTQOtSa7KxGNfauA3R."
                    },
                    "firstName": {
                      "type": "string",
                      "example": "John"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "James"
                    },
                    "email": {
                      "type": "string",
                      "example": "john@email.com"
                    },
                    "credits": {
                      "type": "number",
                      "format": "float",
                      "example": "0"
                    },
                    "isAdmin": {
                      "type": "integer",
                      "description": "User Admin (1 yes, 0 no)",
                      "format": "int32",
                      "example": "false"
                    }
                  },
                  "xml": {
                    "name": "user"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "PUT existing users (Mitarbeiter Verwalten)",
        "description": "This can only be done by the logged in user admin.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "name that need to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update an existent user in the store",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                  },
                  "username": {
                    "type": "string",
                    "example": "theUser"
                  },
                  "password": {
                    "type": "string",
                    "example": "12345"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "James"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@email.com"
                  },
                  "isAdmin": {
                    "type": "integer",
                    "description": "User Admin (1 yes, 0 no)",
                    "format": "int32",
                    "example": 1
                  }
                },
                "xml": {
                  "name": "user"
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "DELETE existing users (Mitarbeiter Verwalten)",
        "description": "This can only be done by the logged in user admin.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}